gif: Fix compiler warning reg. type conversion
With -Wconversion gcc complains with the following warning:

io-gif.c: In function ‘gif_read’:
io-gif.c:246:20: warning: conversion to ‘guint {aka unsigned int}’ from
‘size_t {aka long unsigned int}’ may alter its value [-Wconversion]
    context->ptr += len;
                    ^
io-gif.c:258:28: warning: conversion to ‘guint {aka unsigned int}’ from
‘size_t {aka long unsigned int}’ may alter its value [-Wconversion]
   context->amount_needed = len - (context->size - context->ptr);

This change simply changes their types to match the rhs.

https://bugzilla.gnome.org/show_bug.cgi?id=786261

+++ gdk-pixbuf/io-gif.c
@@ -142,9 +142,9 @@ struct _GifContext
 	GdkPixbufModuleUpdatedFunc update_func;
 	gpointer user_data;
         guchar *buf;
-	guint ptr;
-	guint size;
-	guint amount_needed;
+	gsize ptr;
+	gsize size;
+	gsize amount_needed;
 
 	/* extension context */
 	guchar extension_label;
